Move block querying and archiving in separate module
